feat(cuda_mixed): add path7 full-pcg-fp32 path
extend mixed precision levels with path7 and policy flags

wire path7 through PCG/SpMV interfaces and benchmark all-path tooling

+++ src/backends/cuda_mixed/affine_body/abd_linear_subsystem.cu
@@ -580,7 +580,8 @@ void ABDLinearSubsystem::Impl::retrieve_solution(GlobalLinearSystem::SolutionInf
                 x = info.solution().viewer().name("x")] __device__(int i) mutable
                {
                    // retrieve solution for each body
-                   dq(i) = -x.segment<12>(i * 12).as_eigen();
+                   dq(i) =
+                       (-x.segment<12>(i * 12).as_eigen()).template cast<Float>();
                });
 }
 }  // namespace uipc::backend::cuda_mixed
